Remove ghost hierarchy, require Node on GhostNode, layout always handles ghost nodes, ghostnode system params are private
