--- crates/bevy_ui/src/experimental/mod.rs
@@ -1,17 +1,0 @@
-//! Experimental features are not yet stable and may change or be removed in the future.
-//!
-//! These features are not recommended for production use, but are available to ease experimentation
-//! within Bevy's ecosystem. Please let us know how you are using these features and what you would
-//! like to see improved!
-//!
-//! These may be feature-flagged: check the `Cargo.toml` for `bevy_ui` to see what options
-//! are available.
-//!
-//! # Warning
-//!
-//! Be careful when using these features, especially in concert with third-party crates,
-//! as they may not be fully supported, functional or stable.
-
-mod ghost_hierarchy;
-
-pub use ghost_hierarchy::*;
