+++ crates/bevy_ui/src/lib.rs
@@ -28,9 +28,6 @@ use bevy_derive::{Deref, DerefMut};
 use bevy_picking::PickingSystems;
 use bevy_reflect::{std_traits::ReflectDefault, Reflect};
 mod accessibility;
-// This module is not re-exported, but is instead made public.
-// This is intended to discourage accidental use of the experimental API.
-pub mod experimental;
 mod focus;
 mod geometry;
 mod layout;
