+++ examples/ui/layout/ghost_nodes.rs
@@ -2,17 +2,10 @@
 //!
 //! UI layout will ignore ghost nodes, and treat their children as if they were direct descendants of the first non-ghost ancestor.
 //!
-//! # Warning
-//!
-//! This is an experimental feature, and should be used with caution,
-//! especially in concert with 3rd party plugins or systems that may not be aware of ghost nodes.
-//!
-//! In order to use [`GhostNode`]s you must enable the `ghost_nodes` feature flag.
-
 use bevy::{
     prelude::*,
     text::FontSourceTemplate,
-    ui::experimental::GhostNode,
+    ui::GhostNode,
     ui_widgets::{Activate, Button},
 };
